Removed checked exceptions from lifecycle methods

+++ cas-server-integration-mongo/src/main/java/org/jasig/cas/services/MongoServiceRegistryDao.java
@@ -34,7 +34,7 @@ public final class MongoServiceRegistryDao implements ServiceRegistryDao {
 
     @Autowired
     @NotNull
-    private final MongoOperations mongoTemplate = null;
+    private MongoOperations mongoTemplate;
 
     /**
      * Initialized registry post construction.
@@ -43,17 +43,20 @@ public final class MongoServiceRegistryDao implements ServiceRegistryDao {
      * @throws Exception thrown if collection cant be dropped/created.
      */
     @PostConstruct
-    public void afterPropertiesSet() throws Exception {
-        if (this.dropCollection) {
-            LOGGER.debug("Dropping database collection: {}", this.collectionName);
-            this.mongoTemplate.dropCollection(this.collectionName);
-        }
+    public void afterPropertiesSet() {
+        if (this.mongoTemplate == null) {
+            LOGGER.warn("Mongo template is not correctly configured");
+        } else {
+            if (this.dropCollection) {
+                LOGGER.debug("Dropping database collection: {}", this.collectionName);
+                this.mongoTemplate.dropCollection(this.collectionName);
+            }
 
-        if (!this.mongoTemplate.collectionExists(this.collectionName)) {
-            LOGGER.debug("Creating database collection: {}", this.collectionName);
-            this.mongoTemplate.createCollection(this.collectionName);
+            if (!this.mongoTemplate.collectionExists(this.collectionName)) {
+                LOGGER.debug("Creating database collection: {}", this.collectionName);
+                this.mongoTemplate.createCollection(this.collectionName);
+            }
         }
-
 
     }
